Update date is the Date type now
