+++ stepik-java-api/src/main/java/org/stepik/api/objects/lessons/Lesson.java
@@ -7,8 +7,12 @@ import org.stepik.api.objects.StudyObject;
 import org.stepik.api.urls.Urls;
 
 import java.util.ArrayList;
+import java.util.Date;
 import java.util.List;
 import java.util.Map;
+
+import static org.stepik.api.Utils.timeISOFormat;
+import static org.stepik.api.Utils.toDate;
 
 /**
  * @author meanmail
@@ -48,6 +52,7 @@ public class Lesson extends StudyObject {
     private String createDate;
     @SerializedName("update_date")
     private String updateDate;
+    private transient Date utcUpdateDate;
     @SerializedName("learners_group")
     private String learnersGroup;
     @SerializedName("testers_group")
@@ -386,15 +391,16 @@ public class Lesson extends StudyObject {
     }
 
     @NotNull
-    public String getUpdateDate() {
-        if (updateDate == null) {
-            updateDate = "";
+    public Date getUpdateDate() {
+        if (utcUpdateDate == null) {
+            utcUpdateDate = toDate(updateDate);
         }
-        return updateDate;
+        return utcUpdateDate;
     }
 
-    public void setUpdateDate(@Nullable String updateDate) {
-        this.updateDate = updateDate;
+    public void setUpdateDate(@Nullable Date updateDate) {
+        this.updateDate = timeISOFormat.format(updateDate);
+        utcUpdateDate = updateDate;
     }
 
     @Nullable
