+++ stepik-java-api/src/main/java/org/stepik/api/objects/steps/Step.java
@@ -7,9 +7,13 @@ import org.stepik.api.objects.StudyObject;
 import org.stepik.api.urls.Urls;
 
 import java.util.ArrayList;
+import java.util.Date;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+
+import static org.stepik.api.Utils.timeISOFormat;
+import static org.stepik.api.Utils.toDate;
 
 /**
  * @author meanmail
@@ -45,6 +49,7 @@ public class Step extends StudyObject {
     private String createDate;
     @SerializedName("update_date")
     private String updateDate;
+    private transient Date utcUpdateDate;
     @SerializedName("discussions_count")
     private int discussionsCount;
     @SerializedName("discussion_proxy")
@@ -294,15 +299,16 @@ public class Step extends StudyObject {
     }
 
     @NotNull
-    public String getUpdateDate() {
-        if (updateDate == null) {
-            updateDate = "";
+    public Date getUpdateDate() {
+        if (utcUpdateDate == null) {
+            utcUpdateDate = toDate(updateDate);
         }
-        return updateDate;
+        return utcUpdateDate;
     }
 
-    public void setUpdateDate(@Nullable String updateDate) {
-        this.updateDate = updateDate;
+    public void setUpdateDate(@Nullable Date updateDate) {
+        this.updateDate = timeISOFormat.format(updateDate);
+        utcUpdateDate = updateDate;
     }
 
     public int getDiscussionsCount() {
